+++ workspaces/firehydrant/plugins/firehydrant/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-firehydrant
 
+## 0.18.0
+
+### Minor Changes
+
+- 57ecbb7: Backstage version bump to v1.49.2
+
 ## 0.17.0
 
 ### Minor Changes