+++ workspaces/playlist/plugins/playlist-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-playlist-common
 
+## 0.16.0
+
+### Minor Changes
+
+- d60015e: Backstage version bump to v1.49.2
+
 ## 0.15.0
 
 ### Minor Changes